api_fetch/requirements.txt
backports.csv==1.0.7
cairocffi==1.0.2
CairoSVG==2.4.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
cssselect2==0.2.1
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.0.12
django-export-csv==0.1.3.1
django-import-export==1.2.0
django-queryset-csv==1.0.2
djangorestframework==3.10.0
et-xmlfile==1.0.1
Flask==1.0.3
Flask-SQLAlchemy==2.4.0
html5lib==1.0.1
httpie==1.0.3
idna==2.8
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
MarkupSafe==1.1.1
odfpy==1.4.0
openpyxl==2.6.2
Pillow==6.1.0
pycparser==2.19
Pygments==2.4.2
Pyphen==0.9.5
pytz==2018.9
PyYAML==5.1.1
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.5
tablib==0.13.0
tinycss2==1.0.2
unicodecsv==0.14.1
urllib3==1.25.3
WeasyPrint==48
webencodings==0.5.1
Werkzeug==0.15.4
xlrd==1.2.0
xlwt==1.3.0
